requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
distlib==0.3.0
Django==2.2
django-extensions==2.1.9
django-filter==2.2.0
django-isbn-field==0.5.2
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-jsonapi==3.1.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.10.0
drf-yasg==1.6.1
filelock==3.0.12
flex==6.12.0
future==0.18.2
gunicorn==19.6.0
idna==2.9
inflection==0.3.1
ipython==7.13.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.4
psycopg2==2.8.4
py==1.8.1
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.16.0
python-stdnum==1.13
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
rest-condition==1.0.3
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
tox==2.6.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
validate-email==1.3
virtualenv==20.0.13
wcwidth==0.1.9
Werkzeug==1.0.0